I am

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 /**
-*main - a C program that prints the size of various types on the computer it is compiled and run on.
+*main - a C program that prints the size of various types on
+the computer it is compiled and run on.
 *Return: 0 (success)
 */
 int main(void)
@@ -10,10 +11,10 @@ int n;
 long int l;
 long long int g;
 float f;
-printf("Size of a char: %zu byte(s)\n", sizeof(x));
-printf("size of an int: %zu byte(s)\n", sizeof(n));
-printf("Size of a long int: %zu byte(s)\n", sizeof(l));
-printf("Size of a long long int: %zu byte(s)\n", sizeof(g));
-printf("size of a float: %zu byte(s)\n", sizeof(f));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(x));
+printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(n));
+printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(l));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(g));
+printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
